fix: adapt backend for Vercel serverless
