+++ server/api/index.js
@@ -1,0 +1,4 @@
+const app = require('../app');
+
+// Export pour Vercel Serverless
+module.exports = app;
