+++ server/package.json
@@ -4,7 +4,10 @@
   "description": "'Blog App in MERN STACK '",
   "main": "server.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "start": "node server.js",
+    "dev": "nodemon server.js",
+    "test": "echo \"Error: no test specified\" && exit 1",
+    "prod": "NODE_ENV=production node server.js"
   },
   "author": "",
   "license": "ISC",
@@ -28,6 +31,9 @@
     "multer-s3": "^3.0.1",
     "sharp": "^0.33.5",
     "socket.io": "^4.7.5",
-    "xss": "^1.0.15"
+    "redis": "^4.6.10",
+    "winston": "^3.11.0",
+    "xss": "^1.0.15",
+    "compression": "^1.7.4"
   }
 }
